requirements.txt
streamlit
google-genai
python-dotenv
streamlit-tags
altair
pandas
requests
black
isort